Moved some methods to RoleService and fixed linking for new role and resources not being created.
